// src/main/core/contract/enums/http-status-code.ts
/*
 *
 *  * Copyright (c) 2018, salesforce.com, inc.
 *  * All rights reserved.
 *  * SPDX-License-Identifier: BSD-3-Clause
 *  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
 *
 */

export enum HttpStatusCode {
    SUCCESS = '200',
    BAD_REQUEST = '400',
    UNAUTHORIZED = '401',
    INTERNAL_SERVER_ERROR = '500',
}